compliant, no compile errors nor warnings

+++ src/xmlwriter.cpp
@@ -39,7 +39,7 @@ void XmlWriter::buildActuateContext(std::vector<double> &status){
     using namespace rapidxml;
     function_node = doc->allocate_node(node_element, "function", "actuateBool");
     context_node = doc->allocate_node(node_element, "context", 0);
-    for (int i = 0; i < status.size(); i++){
+    for (size_t i = 0; i < status.size(); i++){
         xml_node<> *dataNode = doc->allocate_node(node_element, "data" + i, std::to_string(std::round(status[i])).c_str());
         context_node->append_node(dataNode);
     }
